chore: mv tqdm test to `test_logging.py`
