+++ tests/test_logging.py
@@ -14,10 +14,12 @@
 
 import os
 import unittest
+from unittest.mock import patch
 
 import transformers.models.bart.tokenization_bart
-from transformers import logging
+from transformers import AutoConfig, logging
 from transformers.testing_utils import CaptureLogger, mockenv, mockenv_context
+from transformers.utils.tqdm_utils import set_progress_bar_enabled
 
 
 class HfArgumentParserTest(unittest.TestCase):
@@ -121,3 +123,17 @@ class HfArgumentParserTest(unittest.TestCase):
             with CaptureLogger(logger) as cl:
                 logger.warning_advice(msg)
             self.assertEqual(cl.out, msg + "\n")
+
+
+def test_set_progress_bar_enabled():
+    TINY_MODEL = "hf-internal-testing/tiny-random-distilbert"
+    with patch("tqdm.auto.tqdm") as mock_tqdm:
+        set_progress_bar_enabled(True)
+        _ = AutoConfig.from_pretrained(TINY_MODEL, force_download=True)
+        mock_tqdm.assert_called()
+
+        mock_tqdm.reset_mock()
+
+        set_progress_bar_enabled(False)
+        _ = AutoConfig.from_pretrained(TINY_MODEL, force_download=True)
+        mock_tqdm.assert_not_called()
